Decode detailed HSM errors from responses
The HSM sends us information about errors which occur in response to
commands which was previously not being decoded.

This commit decodes that information and surfaces it through the
SessionErrorKind enum.
