+++ src/securechannel/mod.rs
@@ -46,4 +46,4 @@ pub use self::context::{Context, CONTEXT_SIZE};
 pub use self::cryptogram::{Cryptogram, CRYPTOGRAM_SIZE};
 pub use self::error::{SecureChannelError, SecureChannelErrorKind};
 pub(crate) use self::mac::{Mac, MAC_SIZE};
-pub(crate) use self::response::{ResponseCode, ResponseMessage};
+pub(crate) use self::response::ResponseMessage;
